==> 05-vagina-vulva-comparison.Rmd <==
---
title: "Comparisons of girls' vaginal and vulvar microbiota"
author: "Roxana J. Hickey <roxana.hickey@gmail.com>"
date: "Last updated March 9, 2015"
output:
  html_document:
    keep_md: yes
    toc: yes
  pdf_document:
    toc: yes
---

***
## Description
This is a supplement to the paper "Vaginal microbiota of adolescent girls prior to the onset of menarche resemble those of reproductive-age women" by Hickey et al. The code works through comparisons of vaginal and vulvar microbiota of perimenarcheal girls. The analyses can be run directly from the R Markdown file using RStudio. It should be run after "01-data-prep.Rmd".

See the project repository at http://github.com/roxanahickey/adolescent for more information.

**Update 2015-01-08: I added “echo=FALSE” options to the chunks of code that make a graph. View full code in R Markdown script.**

**Update 2015-03-09: Updated all figure legends and numbering to be consistent with final publication version (accepted 2015-02-18).**

## Objective
In earlier analyses ("02-hclust-pcoa.Rmd", "03-community-dynamics.Rmd", "04-lmm-lab-ph.Rmd") we focused on vaginal microbiota; now we want to extend our analyses to compare vulvar microbiota alongside vaginal microbiota. Here we perform hierarchical clustering and PCoA with both vaginal and vulvar data (similar to in "02-hclust-pcoa.Rmd") in addition to correlation analysis (part III) and indicator species analysis (part IV).

***
# Initial setup

Clear the workspace; load data and necessary packages.
```{r setup}
rm(list=ls())

load("data-postproc/01-data-prep-last-run.RData")

library(ape)
library(cluster)
library(ggplot2)
library(indicspecies)
library(plyr)
library(reshape)
library(vegan)

## Display session info
sessionInfo()
```

***
# I. Hierarchical clustering of vaginal and vulvar microbiota
The first set of analyses involves clustering vaginal and vulvar samples based on bacterial community composition and selecting the optimal clustering model and number of clusters. The approaches are based on those outlined in the following texts:  
* Legendre P, Legendre L. (2012). _Cluster analysis._ 3rd ed. Elsevier.
* Borcard D, Gillet F, Legendre P. (2011). _Numerical ecology with R._ Springer.

## Hellinger standardization of taxon abundance data
First, we standardize our bacterial read abundances using the Hellinger method. This is a recommended approach when the "species" are sparsely populated at some sites, resulting in many zeros in the species abundance matrix. We then compute the Bray-Curtis dissimilarity matrix from the Hellinger standardized abundance matrix. This will be used in subsequent clustering and ordination analyses.

```{r hellinger-bray-curtis}
## Transform the abundance and proportion matrices because sites should be in rows, species in columns
spe.abund <- t(abund.red)
spe.prop <- t(prop.red)

## Perform the Hellinger transformation
spe.hel <- decostand(spe.abund, method="hellinger")

## Compute the Bray-Curtis dissimilarity matrix from the Hellinger standardized species table
spe.hel.bc <- vegdist(spe.hel, method="bray")
```

## Hierarchical clustering
Now we will perform hierarchical clustering using multiple linkage strategies (single, complete, average/UPGMA, Ward) and select the optimal one using Gower's distance. Then we'll select the optimal number of clusters using the maximum silhouette width.

```{r hclust-comparison, fig.width=12, fig.height=8, echo=FALSE}
## Compute hierarchical clustering using four linkage methods
spe.hb.single <- hclust(spe.hel.bc, method="single")
spe.hb.complete <- hclust(spe.hel.bc, method="complete")
spe.hb.upgma <- hclust(spe.hel.bc, method="average")
spe.hb.ward <- hclust(spe.hel.bc, method="ward.D")

## Plot to compare (these are ugly because the labels obscure each another, 
## but it's useful to take a look at the shape of the dendrograms)
par(mfrow=c(2,2))
plot(spe.hb.single, main="Single Linkage, Hellinger/Bray")
plot(spe.hb.complete, main="Complete Linkage, Hellinger/Bray")
plot(spe.hb.upgma, main="UPGMA, Hellinger/Bray")
plot(spe.hb.ward, main="Ward, Hellinger/Bray")
dev.off()
```

## Selection of optimal clustering model
Next we select the best clustering method by determining the cophenetic distance of each hierarchical clustering, followed by calculation of the Gower distance (Gower 1983), which is the sum of squared differences between the original and cophenetic distances. The method with the lowest Gower distance is considered the optimal clustering model for the distance matrix used. Below, this method identifies average/UPGMA is the best clustering model.

```{r cophenetic}
## Calculate cophenetic distance for each hclust object
spe.hb.single.coph <- cophenetic(spe.hb.single)
spe.hb.complete.coph <- cophenetic(spe.hb.complete)
spe.hb.upgma.coph <- cophenetic(spe.hb.upgma)
spe.hb.ward.coph <- cophenetic(spe.hb.ward)

## Calculate the Gower distance
gow.dist.single <- sum((spe.hel.bc-spe.hb.single.coph)^2)
gow.dist.complete <- sum((spe.hel.bc-spe.hb.complete.coph)^2)
gow.dist.upgma <- sum((spe.hel.bc-spe.hb.upgma.coph)^2)
gow.dist.ward <- sum((spe.hel.bc-spe.hb.ward.coph)^2)

## Compare Gower distances and identify the lowest
gow.dist.single; gow.dist.complete; gow.dist.upgma; gow.dist.ward # Lowest score is UPGMA
```

## Selection of optimal number of clusters
Now we pick the optimal number of clusters according silhouette widths (Rousseew quality index). To do this we plot the average silhouette widths for all partitions except for the trivial partition in a single group (k=1). We find that six clusters are identified as optimal.

```{r hclust-silhouette, fig.width=8, fig.height=6, echo=FALSE}
## First, create an empty vector in which the asw values will be written
asw <- numeric(nrow(spe.abund))

## This function calculates and plots the silhouette width, indicating the optimal number in red (from Borcard et al. 2011)
for (k in 2:(nrow(spe.abund)-1)) {
  sil <- silhouette(cutree(spe.hb.upgma, k=k), spe.hel.bc)
  asw[k] <- summary(sil)$avg.width
}
k.best <- which.max(asw)
plot(1:nrow(spe.abund), asw, type="h", 
     main="Silhouette-optimal number of clusters, UPGMA",
     xlab="k (number of groups)", ylab="Average silhouette width")
axis(1, k.best, paste("optimum",k.best,sep="\n"), col="red", font=2, col.axis="red")
points(k.best, max(asw), pch=16, col="red", cex=1.5)
cat("", "Silhouette-optimal number of clusters k =", k.best, "\n", 
    "with an average silhouette width of", max(asw), "\n")

## Now we can look at how well the number of clusters matches the clustering of our samples:
## Set the number of clusters identified as optimal
k <- 7

## Cut the tree and assign samples to each of the six groups
cutg <- cutree(spe.hb.upgma, k=k)
sil <- silhouette(cutg, spe.hel.bc)
rownames(sil) <- row.names(spe.abund)

## Plot silhouette partition
plot(sil, main="Silhouette plot - Hellinger - UPGMA", cex.names=0.8, col=2:(k+1), nmax=100)

## Plot dendrogram with group labels
hcoplot(spe.hb.upgma, spe.hel.bc, k=6)
```

Assign the group IDs as a new variable in the metadata and define colors for each group:
```{r hclust-newcolors, fig.width=8, fig.height=6}
## Add the group assignments as a variable to the metadata
meta$hclust <- cutg

## Define new colors and names for hclust groups (I determined these by
## looking at the heatmap in the next step then coming back to name the
## clusters and set colors that match the dominant taxon, if any)
col.hclust <- c("1"="slategrey", 
                "2"=col.taxa["Lactobacillus_iners"], 
                "3"=col.taxa["Gardnerella_vaginalis"], 
                "4"=col.taxa["Lactobacillus_gasseri"], 
                "5"=col.taxa["Lactobacillus_crispatus"], 
                "6"=col.taxa["Lactobacillus_jensenii"], 
                "7"=col.taxa["Bifidobacterium"])
names(col.hclust) <- c("Other", "LI", "GV", "LG", "LC", "LJ", "Bifido")

## Replace cluster numbers with new names in metadata
meta$hclust <- gsub("1", "Other", meta$hclust)
meta$hclust <- gsub("2", "LI", meta$hclust)
meta$hclust <- gsub("3", "GV", meta$hclust)
meta$hclust <- gsub("4", "LG", meta$hclust)
meta$hclust <- gsub("5", "LC", meta$hclust)
meta$hclust <- gsub("6", "LJ", meta$hclust)
meta$hclust <- gsub("7", "Bifido", meta$hclust)

## Add these colors as a new variable to col.meta.vag
col.meta$hclust <- col.hclust[meta$hclust]

## Print pie charts to show color palette
par(mar=c(1,1,2,1))
pie(rep(1,7), col=col.hclust, 
    main="Palette: col.hclust\n(with labels)", 
    labels=names(col.hclust))
dev.off()
```

Now we visualize community composition as a heatmap along with the UPGMA dendrogram and cluster assignments just determined:

## Figure S6. Bacterial community composition of the vulvar and vaginal microbiota of girls and vaginal microbiota of mothers sampled longitudinally.
**Note: In the final publication version, I flipped the orientation of the dendrogram horizontally and moved all legend items to the top left of the plot.**

Each column in the dendrogram and heatmap represents the vulvar or vaginal microbiota sampled from a single individual at a single point in time. In total 456 samples are represented: 198 vaginal samples and 211 vulvar samples from 31 girls, and 47 vaginal samples from 24 mothers. e dendrogram represents the average linkage (UPGMA) hierarchical clustering of samples based on the Bray-Curtis dissimilarity matrix computed from Hellinger standardized taxon abundance data. The colored bars below the dendrogram represent sample type (girl/mother, vagina/vulva) and hierarchical cluster assignments. Clusters are named to signify the most abundant taxon, when applicable: LC (_Lactobacillus crispatus_ dominant, n=134), ‘Other’ (n=117), LI (_L. iners_, n=107), LG (_L. gasseri_, n=49), GV (_Gardnerella vaginalis_, n=47), and ‘Bifido’ (_Bifidobacterium_, n=3). e heatmap represents proportions (prior to Hellinger standardization) of the 25 overall most abundant taxa within each community as indicated by the legend at top right.

```{r fig-s6-hclust-heatmap-hellinger-bray-upgma, fig.width=12, fig.height=8, echo=FALSE}
## Sort taxa by abundance (row-wise, by taxon)
rsum <- rowSums(prop.red)

## Pick the top 25 taxa across all samples for the heatmap
pick <- order(rsum,decreasing=TRUE)[1:25]

## Plot the dendrogram and heatmap with selected metadata (uncomment the first line below to save as a PDF)
# pdf("supplemental/fig-s6-hclust-heatmap-hellinger-bray-upgma.pdf", width=12, height=8, pointsize=10)
fcol <- cbind(col.meta$site, col.meta$type, col.hclust[cutg])
colnames(fcol) <- c("Site", "Girl/Mother", "Cluster Group")
par(oma=c(2,2,2,2), lwd=0.75)
heatmap.3(prop.red[pick,],
          col=col.heatmap,
          distfun=vegdist, 
          hclustfun=function (z) hclust(spe.hel.bc, method="average"),
          labCol="",
          ColSideColors=fcol,
          cexRow=1,
          cexCol=0.5,
          mar=c(5,5),
          dendrogram="column",
          Rowv=F,
          Colv=T,
          keysize=1,
          trace="none",
          key=T,
          density.info="none",family="sans")
par(oma=c(1,2,1,2), new=TRUE, xpd=TRUE)
plot(0:1, 0:1, type = "n", axes = F, xlab="", ylab="")
legend(0.15, -0.085, legend=sort(names(col.hclust)), title="Cluster Group", 
       fill=col.hclust[order(names(col.hclust))], bty="n", 
       bg="#ffffff55", inset=0, cex=0.8, ncol=3)
legend(0.4, -0.085, legend=c("Girl", "Mother", "Vagina", "Vulva"), title="Sample Type", 
       fill=c(col.type, col.site), bty="n", bg="#ffffff55", inset=0, cex=0.8, ncol=2)
dev.off()
```

***
# II. PCoA of vaginal and vulvar microbiota
Now we perform PCoA to obtain a more nuanced picture of the similarities and differences among samples. Again, the approaches are based on those outlined in the following texts:  
* Legendre P, Legendre L. (2012). _Cluster analysis_. 3rd ed. Elsevier.
* Borcard D, Gillet F, Legendre P. (2011). _Numerical ecology with R_. Springer.

## Setup PCoA
*Note: the PCoA method below produces negative eigenvalues unless corrected, which can be problematic for interpreting the R^2\-like ratio (essentially variance explained by an eigenvalue in PCA). See Legendre & Legendre Numerical Ecology Ch 9 for more discussion of this (p. 505 in 3rd edition 2012). However, as long as the largest-value negative eigenvalue is smaller in absolute value than any of the positive eigenvalues of interest (typically the first two), the interpretation is still meaningful. A correction was suggested by Cailliez & Pagès to adjust the R^2\-like ratio when negative eigenvalues are present -- see Legendre & Legendre p. 506, eq. 9.48.*

```{r pcoa-setup}
## Calculate PCoA on Bray-Curtis dissimilarity matrix
spe.hb.pcoa <- cmdscale(vegdist(spe.hel), eig=TRUE, k=nrow(spe.hel)-1)

## Calculate species scores
spe.hb.wa <- wascores(spe.hb.pcoa$points, spe.hel)

## Apply Cailliez correction using ape::pcoa to obtain R^2-like ratios
spe.hb.c.pcoa <- pcoa(vegdist(spe.hel), correction="cailliez")

## R^2-like ratio for first three axes
spe.hb.c.pcoa$values$Rel_corr_eig[1:3]

## ~ Total variance explained by first two axes
sum(spe.hb.c.pcoa$values$Rel_corr_eig[1:2])

## Note: we will still use the PCoA computed from cmdscale for plotting 
## since it is compatible with vegan's ordiplot functions (next section). 
## I had difficulty applying the Cailliez correction to the cmdscale PCoA 
## object which is why I used the above ape::pcoa function. If you run it 
## without the Cailliez correction (uncomment spe.hb.pcoa.2 above first), 
## you get the same eigenvalues as from the cmdscale method. See below:

## PCoA from ape::pcoa without the Cailliez correction
spe.hb.pcoa.2 <- pcoa(vegdist(spe.hel))

## Compare to PCoA from stats::cmdscale
head(spe.hb.pcoa$eig)
head(spe.hb.pcoa.2$values$Eigenvalues)
```

## PCoA plots
Now we can plot the PCoA and overlay different variables with point shapes and colors. We'll keep it simple by using only the first 2 axes, which as we saw above account for ~22% of the variance after applying the Cailliez correction. We'll look at the same plot colored according to different metadata variables to look for any interesting patterns in the data:

* Color-coding by hierarchical cluster group
* Color-coding by body site

```{r pcoa-vag-vul-hclust, fig.width=7, fig.height=7, echo=FALSE}
## Make semi-transparent colors for plotting 3D scatterplot
col.hclust.alpha70 <- makeTransparent(col.hclust[meta$hclust], alpha=0.70)
col.site.alpha70 <- makeTransparent(col.site[meta$site], alpha=0.70)
col.type.alpha70 <- makeTransparent(col.type[meta$type], alpha=0.70)

## Plot the first two axes of the PCoA
# pdf("misc/pcoa-girl-vag-vul-hclust.pdf", width=6, height=6, pointsize=10)
ordiplot(scores(spe.hb.pcoa)[,c(1,2)], type="n", xlab="Axis 1", ylab="Axis 2")
abline(h=0, lty=3, col="gray70")
abline(v=0, lty=3, col="gray70")
points(scores(spe.hb.pcoa), 
       col=col.hclust.alpha70, 
       pch=col.meta$pch.sample.gp,
       cex=1.3)
legend("topright", bty="n", title="Sample Type",
       legend=c("girl vag pre", "girl vag post", "girl vag NA", 
                "girl vul pre", "girl vul post", "girl vul NA", 
                "mom vag"), 
       pch=c(1,16,10,0,15,12,17))
legend("bottomright", bty="n", bg="white", legend=sort(names(col.hclust)), 
       title="Cluster Group", fill=col.hclust[order(names(col.hclust))])
# dev.off()

## Colored by body site
# pdf("misc/pcoa-girl-vag-vul-site.pdf", width=6, height=6, pointsize=10)
ordiplot(scores(spe.hb.pcoa)[,c(1,2)], type="n", xlab="Axis 1", ylab="Axis 2")
abline(h=0, lty=3, col="gray70")
abline(v=0, lty=3, col="gray70")
points(scores(spe.hb.pcoa)[meta$type=="mom",c(1,2)], 
       col=col.type.alpha70[meta$type=="mom"], 
       pch=col.meta$pch.sample.gp[meta$type=="mom"],
       cex=1.3)
points(scores(spe.hb.pcoa)[meta$type=="girl",c(1,2)], 
       col=col.site.alpha70[meta$type=="girl"], 
       pch=col.meta$pch.sample.gp[meta$type=="girl"],
       cex=1.3)
legend("topright", bty="n", title="Sample Type",
       legend=c("girl vag pre", "girl vag post", "girl vag NA", 
                "girl vul pre", "girl vul post", "girl vul NA", 
                "mom vag"), 
       pch=c(1,16,10,0,15,12,17), col=c(rep("black", 6), col.type[2]))
legend("bottomright", bty="n", legend=c("vag", "vul"), title="Site", fill=col.site[c("vag","vul")])
# dev.off()
```

***
# III. Correlation of paired vaginal and vulvar samples
Next, we want to determine correlations among paired vagina-vulva samples to get an idea of how similar they are in taxonomic composition. Pearson correlation is one of the most commonly used measures and calculates the strength of the linear relationship between normally distributed variables. However, our variables (taxon abundances) are not expected to be related linearly, and in most cases they are not normally distributed. Outliers or extreme values therefore have a strong influence on the correlation. Spearman correlation, on the other hand, considers only the rank-order of variables. Spearman makes no assumptions about the distributions or linear relationship among variables, so large and small outliers have much less effect. This is more appropriate to our dataset. Because we are considering rank-order only, we would get the same result on either the taxon proportion table or the Hellinger-standardized abundance table. Below we calculate the Spearman correlation on the untranformed taxon proportion table. 

```{r vag-vul-spearman, fig.width=8, fig.height=6}
meta.paired <- subset(meta.orig, vag==TRUE & vul==TRUE & type=="girl")
prop.vag.pr <- prop.red[,colnames(prop.red) %in% meta.paired$sample.ID.vag]
prop.vul.pr <- prop.red[,colnames(prop.red) %in% meta.paired$sample.ID.vul]

## Compute Spearman correlation on proportion data
cor.vag.vul.sp <- cor(prop.vag.pr, prop.vul.pr, method="spearman")

## Extract diagonal (vagina-vulva pairs)
cor.vag.vul.sp.pr <- diag(cor.vag.vul.sp)
hist(cor.vag.vul.sp.pr, breaks=30)
mean(cor.vag.vul.sp.pr)
median(cor.vag.vul.sp.pr)

meta.paired$cor.spearman <- cor.vag.vul.sp.pr

gg.cor.vag.vul.sp <- ggplot(meta.paired, aes(x=visit, y=cor.spearman, shape=men.stat))

gg.cor.vag.vul.sp + geom_point(size=3) + 
  geom_line(lty=3) +
  facet_wrap( ~ subject, ncol=4) +
  scale_shape_manual(values=c(16, 1), 
                     breaks=c("pre", "post"), 
                     labels=c("Pre", "Post"),
                     name="Menarche\nStatus", na.value=10) +
  xlab("Visit") +
  ylab("Spearman's rho of vagina vs. vulva") +
  scale_x_discrete(1:14, name="Visit No.") +
  ylim(0,1) +
  theme_cust_nominor +
  theme(axis.text=element_text(size=6),
        legend.position=c(1,0),
        legend.justification=c(1,0))
```

For curiosity's sake we can also compute the Pearson correlations (better to do so on the Hellinger-standardized abundance data so as to lessen effects of very high and very low abundances). The correlations are indeed higher for most cases except those with the starkest differences in composition. However, this should be interpreted with caution since Pearson assumes normal distributions and linear dependence among variables, and many taxa in our dataset have heavily skewed abundance distributions even after the Hellinger transformation (the Spearman method does not rely on these assumptions).

```{r vag-vul-pearson-setup}
spe.hel <- decostand(t(abund.red), method="hellinger")
spe.hel.vag.pr <- t(spe.hel[rownames(spe.hel) %in% meta.paired$sample.ID.vag,])
spe.hel.vul.pr <- t(spe.hel[rownames(spe.hel) %in% meta.paired$sample.ID.vul,])

## Compute the Pearson correlation on Hellinger-transformed data
cor.vag.vul.hel.pe <- cor(spe.hel.vag.pr, spe.hel.vul.pr, method="pearson")

## Extract diagonal (vagina-vulva pairs)
cor.vag.vul.hel.pe.pr <- diag(cor.vag.vul.hel.pe)
hist(cor.vag.vul.hel.pe.pr, breaks=30)
mean(cor.vag.vul.hel.pe.pr)
median(cor.vag.vul.hel.pe.pr)

meta.paired$cor.pearson.hellinger <- cor.vag.vul.hel.pe.pr
```

```{r vag-vul-pearson, fig.width=8, fig.height=6, echo=FALSE}
gg.cor.vag.vul.pe <- ggplot(meta.paired, aes(x=visit, y=cor.pearson.hellinger, shape=men.stat))

gg.cor.vag.vul.pe + geom_point(size=3) + 
  geom_line(lty=3) +
  facet_wrap( ~ subject, ncol=4) +
  scale_shape_manual(values=c(16, 1), 
                     breaks=c("pre", "post"), 
                     labels=c("Pre", "Post"),
                     name="Menarche\nStatus", na.value=10) +
  xlab("Visit") +
  ylab("Pearson correlation of vagina vs. vulva") +
  scale_x_discrete(1:14, name="Visit No.") +
  ylim(0,1) +
  theme_cust_nominor +
  theme(axis.text=element_text(size=6),
        legend.position=c(1,0),
        legend.justification=c(1,0))

## Plot together
cor.meta.pr <- meta.paired[,c("subject", "visit", "men.stat",
                              "cor.spearman", "cor.pearson.hellinger")]
cor.meta.pr <- melt(cor.meta.pr, id.vars=c("subject", "visit", "men.stat"))

gg.cor.vag.vul <- ggplot(cor.meta.pr, aes(x=visit, y=value, shape=men.stat, color=variable))

gg.cor.vag.vul + geom_point(size=3) + 
  geom_line(lty=3) +
  facet_wrap( ~ subject, ncol=4) +
  scale_shape_manual(values=c(16, 1), 
                     breaks=c("pre", "post"), 
                     labels=c("Pre", "Post"),
                     name="Menarche\nStatus", na.value=10) +
  xlab("Visit") +
  ylab("Correlation of vagina vs. vulva") +
  scale_x_discrete(1:14, name="Visit No.") +
  ylim(0,1) +
  theme_cust_nominor +
  theme(axis.text=element_text(size=6))
```

***
# IV. Indicator 'species' (taxa) analysis of vaginal vs. vulvar samples
Indicator species analysis was performed to identify taxa most strongly associated with, but not necessarily most abundant within, different groups of samples. The first analysis was performed on vaginal and vulvar samples from girls only. Samples were separated into four groups encompassing premenarcheal and postmenarcheal vaginal and vulvar samples. Taxa with an indicator value of at least 0.5 and p-value less than or equal to 0.01 are results reported below.

## IndVal by body site and menarcheal status
We first test for indicator "species" (in our case, taxa which include species, genus, family and order levels) ofgroups of samples separated by type:
* Girl vagina premenarche (n=110)
* Girl vagina postmenarche (n=87)
* Girl vulva premenarche (n=119)
* Girl vulva postmenarche (n=91)

*Note: we will ignore the two samples for which menarche status is unknown, 13306Vag1 and 13306Vul1. I chose not to include the mother samples in this analysis because there are substantially fewer samples in that group, and groups LG and Other were very much underrepresented (even though we have commonly observed similar community types in reproductive age women in other studies).*

In this analysis we will test for indicator taxa of combinations of sample groups in addition to single groups. Note that slightly different results may be obtained each time this is run, particularly for the more weakly associated taxa. For that reason, the results reported in the manuscript include only those with an IndVal statistic >=0.5 and p-value <= 0.01.

### Indicator taxa for groups of vaginal and vulvar microbiota of adolescent girls
```{r indic-species-sampletype-groups}
## Reduce the metadata and taxon table for ease of use below
meta.girl <- subset(meta, type=="girl" & men.stat %in% c("pre", "post"))
prop.girl <- prop.red[,meta$type=="girl" & meta$men.stat %in% c("pre","post")]

## Calculate IndVal and summarize (note that slightly different results may be obtained each time)
ind.girl.sampletype <- multipatt(data.frame(t(prop.girl)), meta.girl$type.site.men.stat, func="IndVal.g")
summary(ind.girl.sampletype, alpha=0.01) # list all results with p<=0.01

## Extract results for which IndVal stat >= 0.5 and pval <= 0.01
ind.girl.sampletype.res <- ind.girl.sampletype$sign[ind.girl.sampletype$sign$p.value <= 0.01 & ind.girl.sampletype$sign$stat >= 0.5,]
remove.na <- grep("NA", rownames(ind.girl.sampletype.res))
ind.girl.sampletype.res <- ind.girl.sampletype.res[-remove.na,]

## Run a bootstrapped analysis of IndVal associations for all individual groups
ind.girl.sampletype.boot <- strassoc(data.frame(t(prop.girl)), meta.girl$type.site.men.stat, func="IndVal.g", nboot=100)
rownames(ind.girl.sampletype.boot$stat) <- rownames(ind.girl.sampletype.boot$lowerCI)
rownames(ind.girl.sampletype.boot$stat) <- gsub(" ", ".", rownames(ind.girl.sampletype.boot$stat))
ind.girl.sampletype.boot.assoc <- ind.girl.sampletype.boot$stat[rownames(ind.girl.sampletype.boot$stat) %in% rownames(ind.girl.sampletype.res),]

## Combine IndVal stats/pvals with bootstrapped associations and write table
ind.girl.sampletype.df <- cbind(ind.girl.sampletype.res, 
                                ind.girl.sampletype.boot.assoc[match(rownames(ind.girl.sampletype.res), 
                                                                     rownames(ind.girl.sampletype.boot.assoc)),])
write.csv(ind.girl.sampletype.df, "misc/indicspecies-sampletype-groups.csv")
```

## IndVal for UPGMA hierarchical cluster groups
The second indicator species analysis is based on the six groups determined by hierarchical clustering earlier. This is somewhat circular since those groups were clustered on the basis of community composition, but it does help to highlight which taxa are most strongly associated with each group, which might be difficult to determine just looking at the dendrogram and heatmap. We're considering all combinations of groups here, but you could also look at taxa associated with individual groups by adding the option duleg==TRUE to the multipatt() command.

### Indicator taxa for groups of vaginal and vulvar samples from girls and mothers as determined by UPGMA hierarchical clustering

```{r indic-species-hclust-groups}
## Calculate IndVal and summarize
ind.hbaclust.single <- multipatt(data.frame(t(prop.red)), meta$hclust, func="IndVal.g")
summary(ind.hbaclust.single, alpha=0.01) # list all results with p<=0.01

## Summarize results with p<=0.01
ind.hbaclust.single.res <- ind.hbaclust.single$sign[ind.hbaclust.single$sign$p.value <= 0.01,]
remove.na <- grep("NA", rownames(ind.hbaclust.single.res))
ind.hbaclust.single.res <- ind.hbaclust.single.res[-remove.na,]

## Run a bootstrapped analysis of IndVal associations for all individual groups
ind.hbaclust.single.boot <- strassoc(t(prop.red), meta$hclust, func="IndVal.g", nboot=100)
rownames(ind.hbaclust.single.boot$stat) <- rownames(ind.hbaclust.single.boot$lowerCI)
rownames(ind.hbaclust.single.boot$stat) <- gsub(" ", ".", rownames(ind.hbaclust.single.boot$stat))
ind.hbaclust.single.boot.assoc <- ind.hbaclust.single.boot$stat[rownames(ind.hbaclust.single.boot$stat) %in% rownames(ind.hbaclust.single.res),]

## Combine IndVal stats/pvals with bootstrapped associations and write table
ind.hbaclust.single.df <- cbind(ind.hbaclust.single.res, 
                                ind.hbaclust.single.boot.assoc[match(rownames(ind.hbaclust.single.res), 
                                                                     rownames(ind.hbaclust.single.boot.assoc)),])
write.csv(ind.hbaclust.single.df, "misc/indicspecies-hclust-groups.csv")
```

***
# Save R workspace
This will save the workspace (data) in two separate images: one named with today's date, in case you ever need to restore that version, and another with a non-dated name that can be easily loaded into subsequent analyses.

### Cleanup
```{r cleanup}
rm(fcol, ind.girl.sampletype, ind.girl.sampletype.boot, ind.girl.sampletype.boot.assoc,
   ind.girl.sampletype.res, ind.hbaclust.single, ind.hbaclust.single.boot, 
   ind.hbaclust.single.boot.assoc, ind.hbaclust.single.res, asw, cutg, gg.cor.vag.vul,
   gg.cor.vag.vul.pe, gg.cor.vag.vul.sp, gow.dist.complete, gow.dist.single,
   gow.dist.upgma, gow.dist.ward, k, k.best, pick, remove.na, rsum, sil,
   spe.hb.complete.coph, spe.hb.single.coph, spe.hb.upgma.coph, spe.hb.ward.coph)
```

```{r save-workspace}
save.image(paste("data-postproc/05-vaginal-vulva-comparison-", Sys.Date(), ".RData", sep=""))
save.image(paste("data-postproc/05-vaginal-vulva-comparison-last-run.RData", sep=""))
```
