==> 06-post-review.Rmd <==
---
title: "Post-review additional analyses"
author: "Roxana J. Hickey <roxana.hickey@gmail.com>"
date: "Last updated March 9, 2015"
output:
  html_document:
    keep_md: yes
    toc: yes
  pdf_document:
    toc: yes
---

***
## Description
This is a supplement to the paper "Vaginal microbiota of adolescent girls prior to the onset of menarche resemble those of reproductive-age women" by Hickey et al. These analyses were added after an initial round of peer review. Minor changes were made to other scripts and noted accordingly.

See the project repository at http://github.com/roxanahickey/adolescent for more information.

**Update 2015-03-09: Updated all figure legends and numbering to be consistent with final publication version (accepted 2015-02-18).**

## Objective
These analyses address two reviewer requests: 1) explore how pH varies with respect to community cluster group (from the hierarchical clustering in 02-hclust-pcoa.Rmd) and 2) how community diversity changes with respect to Tanner stage, menarche status and pH.

***
## Initial setup

Load in data from previous analyses.
```{r setup}
rm(list=ls())

load("data-postproc/01-data-prep-last-run.RData")
load("data-postproc/02-hclust-pcoa-last-run.RData")
load("data-postproc/03-community-dynamics-last-run.RData")
load("data-postproc/04-lmm-lab-ph-last-run.RData")
load("data-postproc/05-vaginal-vulva-comparison-last-run.RData")

library(ggplot2)
library(vegan)
library(reshape)
library(tidyr)
library(grid)
library(nparcomp)
library(pgirmess)

source("http://www.jennajacobs.org/R/rarefaction.txt")

## Display session info
sessionInfo()
```

# I. Vaginal pH across community cluster groups
'Bifido' cluster is excluded because there are only two samples, and pH data for only one sample. The second plot will be saved as Figure S2.
```{r ph-cluster-box, fig.width=8, fig.height=6, echo=FALSE}
# colored by girl/mom
gg.ph.hclust <- ggplot(subset(meta.vag, hclust %in% c("GV","LC","LG","LI","LJ","Other")), 
                              aes(y=ph, x=hclust, fill=type))
box.ph.hclust <- gg.ph.hclust + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.type,
                    name="Cluster Group") +
  xlab("Hierarchical cluster group") +
  ylab("Vaginal pH") +
  theme_cust

print(box.ph.hclust)
```

## Figure S2. Vaginal pH across hierarchical cluster groups.
**Note: I annotated the figure with p-values in Illustrator for the final publication version.**

Vaginal microbiota were grouped into hierarchical clusters, listed in the legend at right. A multiple comparisons test using Tukey's method was used to identify significant differences in vaginal pH between clusters. P-values less than 0.05 are shown on the plot with a connector indicating the two groups being compared. Each box represents the interquartile range, the whiskers represent the upper and lower quartiles, the horizontal line represents the median, and open circles represent outliers.
```{r fig-s2-ph-cluster, fig.width=8, fig.height=6, echo=FALSE}
# colored by cluster group, no separation of girls/mothers
gg.ph.hclust2 <- ggplot(subset(meta.vag, hclust %in% c("GV","LC","LG","LI","LJ","Other")), 
                        aes(y=ph, x=hclust, fill=hclust))
box.ph.hclust2 <- gg.ph.hclust2 + 
  geom_boxplot(width=1, outlier.shape=1) +
  scale_fill_manual(values=col.hclust.vag,
                    name="Cluster Group") +
  ylim(4,8) +
  xlab("Cluster Group") +
  ylab("Vaginal pH") +
  theme_cust

print(box.ph.hclust2)
# ggsave("supplemental/fig-s2-ph-cluster.pdf", width=8, height=6, units="in")
```

### Kruskal-Wallis rank sum test
It appears from these plots that vaginal pH is considerably higher in the GV and Other clusters. Now we'll perform some simple stats to see how significant these differences are. We'll use the Kruskal-Wallis rank sum test, a nonparametric group comparison test. We are violating some assumptions because we technically have repeated measures within individuals, but to get a rough idea of differences in pH across groups we're going to treat the groups as independent. We will, however, exclude the Bifido group as it only has 2 samples (and one of them is missing pH data). We'll also exclude the 77 cases for which pH data are unavailable.
```{r ph-kruskal}
summary(factor(meta.vag$hclust))

meta.vag.ph <- na.exclude(meta.vag[,c("subject","type","ph","hclust")]) # n=168
# Remove Bifido group samples
meta.vag.ph <- subset(meta.vag.ph, hclust %in% c("GV","LC","LG","LI","LJ","Other")) # n=167

kruskal.test(ph ~ factor(hclust), data=meta.vag.ph)
```

This test confirms there is a significant difference among the six groups (p=4.7E-08), but not which groups differ from others. The kruskalmc function from package pgirmess allows us to perform multiple pairwise comparisons after a Kruskal-Wallis test (i.e. a posthoc test). According to documentation the procedure was developed by Patrick Giraudoux based on Siegel and Castellan (1988) Non parametric statistics for the behavioural sciences. MacGraw Hill Int., New York. pp 213-214.
```{r ph-kruskal-mc}
kruskalmc(ph ~ factor(hclust), data=meta.vag.ph)
```

### Tukey's test for multiple comparisons
The kruskalmc test above identified the following contrasts as significantly different ("TRUE" in table): GV vs. LC, GV vs. LI, Other vs. LC, Other vs. LI. An alternative and more robust approach that will also give us p-values is the nparcomp function from the package by the same name. We'll use the Tukey method. Note again that we are treating samples as independent although truly they are not.
```{r ph-tukey-mc}
tukey.mc <- nparcomp(ph ~ factor(hclust), data=meta.vag.ph, type = "Tukey")
summary(tukey.mc)
```

The test above identifies the same four contrasts as significant, as well as LC vs. LJ (p=0.001). This confirms our visual observation that vaginal pH is significantly higher in the GV and Other clusters relative to each of the _Lactobacillus_-dominant clusters. Save the results as a table and annotate significant values on Figure S2.
```{r ph-tukey-res}
res.tukey <- data.frame(tukey.mc$Analysis)
x <- gsub("p( ", "", res.tukey$Comparison, fixed=TRUE)
x <- gsub(",", "vs.", x)
x <- gsub(" )", "", x, fixed=TRUE)
res.tukey$Comparison <- x

write.csv(res.tukey, "supplemental/ph-hclust-tukey.csv", quote=FALSE, row.names=FALSE)
```

# II. Diversity analysis

## Extract genus-level abundance data
Previously (in 01-data-prep.Rmd) we summarized a taxon abundance table that included classifications at several levels (species, genus, etc.). For the diversity analysis we want to work at one level at a time. The table 79taxa_noNA.csv matches each taxon name from the original table to its corresponding genus, family, order and phylum when applicable, and lists "Other" if the classification is not available. We'll use this to construct new abundance tables for each classification level.
```{r taxa-levels}
taxa.levels <- read.csv("data/79taxa_noNA.csv", header=TRUE)

# first attempt, better below
# class.sum <- function(ab, lv, rank="genus") {
#   tmp <- cbind(lv[colnames(lv)==rank], ab)
#   abund.tmp <- melt(tmp, id.vars=rank)
#   colnames(abund.tmp)[1] <- "level"
#   assign(paste("abund.", rank, sep=""), cast(abund.tmp, variable~level, sum))
# }
# 
# abund.gen <- class.sum(abund.red, taxa.levels, rank="genus")

# function courtesy of Matt Pennell
class_sum <- function(ab, lv, rank="genus"){
    tax <- lv[,c(rank)]
    lab <- lapply(unique(tax), function(x) which(tax == x))
    ## subset abundance by each
    ab_tax <- lapply(lab, function(x) ab[x,])
    res <- lapply(ab_tax, function(x) colSums(x))
    out <- do.call(rbind, res)
    rownames(out) <- unique(tax)
    out
}

abund.gen <- t(class_sum(abund.red, taxa.levels, rank="genus"))
abund.fam <- t(class_sum(abund.red, taxa.levels, rank="family"))
abund.ord <- t(class_sum(abund.red, taxa.levels, rank="order"))
abund.phy <- t(class_sum(abund.red, taxa.levels, rank="phylum"))
```

## Rarefaction analysis
We know from earlier that there is quite a bit of variance in read count across samples. Therefore we want to rarefy our samples prior to doing diversity analyses. Below is a summary of read count number across all 456 samples.
```{r read-counts}
summary(rowSums(abund.gen))

hist(rowSums(abund.gen), breaks=50, col="yellow2", xlab="Number of Reads", 
     main="Sample Read Counts")
```

Here are genus-level rarefaction curves for all 456 samples. Note that in each case, 'Other' is included as its own category, even though it encompasses many taxa that could not be classified to a given level (applies mainly to the genus and family levels). Lines represent the observed number of genera, families, orders or phyla for each vaginal or vulvar swab sample. The x-axis extends to 5000 samples (reads), although the maximum read count was 31,570.
```{r rare-curves-all, fig.width=12, fig.height=8, echo=FALSE}
par(mfrow=c(2,2))

# Genus
rarecurve(abund.gen[rownames(abund.gen) %in% rownames(meta.vag),], 
          col=col.site.alpha70["vag"], step=100, xlim=c(0,5000), ylim=c(0,50), 
          main="Genus", label=FALSE, xlab="No. reads sampled", ylab="No. genera observed")
par(new=TRUE)
rarecurve(abund.gen[rownames(abund.gen) %in% colnames(prop.vul.pr),],
          col=col.site.alpha70["vul"], step=100, xlim=c(0,5000), ylim=c(0,50),
          main="", label=FALSE, xlab="", ylab="")
legend("topright", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n", horiz=TRUE)

# Family
rarecurve(abund.fam[rownames(abund.fam) %in% rownames(meta.vag),], 
          col=col.site.alpha70["vag"], step=100, xlim=c(0,5000), ylim=c(0,30), 
          main="Family", label=FALSE, xlab="No. reads sampled", ylab="No. families observed")
par(new=TRUE)
rarecurve(abund.fam[rownames(abund.fam) %in% colnames(prop.vul.pr),],
          col=col.site.alpha70["vul"], step=100, xlim=c(0,5000), ylim=c(0,30),
          main="", label=FALSE, xlab="", ylab="")
legend("topright", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n", horiz=TRUE)

# Order
rarecurve(abund.ord[rownames(abund.ord) %in% rownames(meta.vag),], 
          col=col.site.alpha70["vag"], step=100, xlim=c(0,5000), ylim=c(0,20), 
          main="Order", label=FALSE, xlab="No. reads sampled", ylab="No. orders observed")
par(new=TRUE)
rarecurve(abund.ord[rownames(abund.ord) %in% colnames(prop.vul.pr),],
          col=col.site.alpha70["vul"], step=100, xlim=c(0,5000), ylim=c(0,20),
          main="", label=FALSE, xlab="", ylab="")
legend("topright", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n", horiz=TRUE)

# Phylum
rarecurve(abund.phy[rownames(abund.phy) %in% rownames(meta.vag),], 
          col=col.site.alpha70["vag"], step=100, xlim=c(0,5000), ylim=c(0,10), 
          main="Phylum", label=FALSE, xlab="No. reads sampled", ylab="No. phyla observed")
par(new=TRUE)
rarecurve(abund.phy[rownames(abund.phy) %in% colnames(prop.vul.pr),],
          col=col.site.alpha70["vul"], step=100, xlim=c(0,5000), ylim=c(0,10),
          main="", label=FALSE, xlab="", ylab="")
legend("topright", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n", horiz=TRUE)

dev.off()
```

It's also helpful to check accumulation curves to give an idea of how the number of observed taxa is dependent on our sampling depth (i.e. number of vaginal/vulvar swabs collected). In all cases they level out before 50 samples, which is well below our actual sampling number.
```{r spec-accum, fig.width=12, fig.height=8, echo=FALSE}
par(mfrow=c(2,2))

# Genus
plot(specaccum(abund.gen[rownames(abund.gen) %in% rownames(meta.vag),], 
               method="random", permutations=100, conditioned=FALSE),
     col=col.site.alpha70["vag"], lwd=2, xlim=c(0,250), ylim=c(0,65),
     xlab="No. samples", ylab="Taxa observed", main="Genus")
par(new=TRUE)
plot(specaccum(abund.gen[rownames(abund.gen) %in% colnames(prop.vul.pr),], 
               method="random", permutations=100, conditioned=FALSE), 
     col=col.site.alpha70["vul"], lwd=2, xlim=c(0,250), ylim=c(0,65),
     xlab="", ylab="")
legend("right", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n")

# Family
plot(specaccum(abund.fam[rownames(abund.fam) %in% rownames(meta.vag),], 
               method="random", permutations=100, conditioned=FALSE),
     col=col.site.alpha70["vag"], lwd=2, xlim=c(0,250), ylim=c(0,35),
     xlab="No. samples", ylab="Taxa observed", main="Family")
par(new=TRUE)
plot(specaccum(abund.fam[rownames(abund.fam) %in% colnames(prop.vul.pr),], 
               method="random", permutations=100, conditioned=FALSE), 
     col=col.site.alpha70["vul"], lwd=2, xlim=c(0,250), ylim=c(0,35),
     xlab="", ylab="")
legend("right", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n")

# Order
plot(specaccum(abund.ord[rownames(abund.ord) %in% rownames(meta.vag),], 
               method="random", permutations=100, conditioned=FALSE),
     col=col.site.alpha70["vag"], lwd=2, xlim=c(0,250), ylim=c(0,20),
     xlab="No. samples", ylab="Taxa observed", main="Order")
par(new=TRUE)
plot(specaccum(abund.ord[rownames(abund.ord) %in% colnames(prop.vul.pr),], 
               method="random", permutations=100, conditioned=FALSE), 
     col=col.site.alpha70["vul"], lwd=2, xlim=c(0,250), ylim=c(0,20),
     xlab="", ylab="")
legend("right", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n")

# Phylum
plot(specaccum(abund.phy[rownames(abund.phy) %in% rownames(meta.vag),], 
               method="random", permutations=100, conditioned=FALSE),
     col=col.site.alpha70["vag"], lwd=2, xlim=c(0,250), ylim=c(0,10),
     xlab="No. samples", ylab="Taxa observed", main="Phylum")
par(new=TRUE)
plot(specaccum(abund.phy[rownames(abund.phy) %in% colnames(prop.vul.pr),], 
               method="random", permutations=100, conditioned=FALSE), 
     col=col.site.alpha70["vul"], lwd=2, xlim=c(0,250), ylim=c(0,10),
     xlab="", ylab="")
legend("right", c("Vagina","Vulva"), lty=c(1,1), col=col.site, bty="n")

dev.off()
```

## Calculate diversity indices based on fixed sampling depth
Based on the rarefaction curves, we should be safe to subsample n=2000 at the genus level, as most of the curves are saturated by that stage. This will prevent high-read samples from having inflated diversity estimates. 39 samples have less than 2000 sequences, so rather than sample them with replacement we will exclude them from further analysis. At this point we'll calculate genus richness, obtain rarefied genus abundance/proportion tables, and calculate diversity indices. Again, note that 'Other' is retained as a separate category even though it encompasses several taxa that could not be classified at the genus level. Our estimates will therefore be (probably very) conservative.

```{r rare-setup}
# Reduce dataset to samples with at least 2000 reads (excludes 39 samples)
abund.gen.2000 <- abund.gen[rowSums(abund.gen)>=2000,] # 417 samples

# Reduce metadata table accordingly with above (to be combined in a few steps)
meta.red <- meta[rownames(meta) %in% rownames(abund.gen.2000),]
ord <- sort(rownames(meta.red))

# Calculate genus richness based on subsampling at n=2000
rare.rich.gen <- rarefy(abund.gen.2000, sample=2000)

# Make rarefied genus abundance and proportion tables
rare.abund.gen.2000 <- rrarefy(abund.gen.2000, sample=2000)
rare.prop.gen.2000 <- prop.table(rare.abund.gen.2000, margin=1)  

# Calculate diversity indicies from rarefied proportion data
rare.shan.gen <- diversity(rare.prop.gen.2000, index="shannon")
rare.simp.gen <- diversity(rare.prop.gen.2000, index="simpson")
rare.invsimp.gen <- diversity(rare.prop.gen.2000, index="invsimpson")

# Combine richness and diversity data with previous metadata
meta.red.comb <- cbind(meta.red[ord,], 
                       richness=rare.rich.gen[ord], 
                       shannon=rare.shan.gen[ord],
                       simpson=rare.simp.gen[ord],
                       invsimp=rare.invsimp.gen[ord])
```

Let's have a quick look at the distributions and relationships among genus richness and diversity indices. I chose to test multiple diversity indices, having no strong a priori reason to prefer one over another. Simpson is often preferred over Shannon because Shannon is sensitive to richness differences among samples. Therefore we'll report Simpson diversity along with richness estimates in the supplemental materials, but continue to explore all of the data here.
```{r div-hist, fig.width=12, fig.height=8, echo=FALSE}
par(mfrow=c(2,2))
hist(meta.red.comb$richness, breaks=20, main="Richness", xlab="Value", col="turquoise")
hist(meta.red.comb$shannon, breaks=20, main="Shannon", xlab="Value", col="yellow2")
hist(meta.red.comb$simpson, breaks=20, main="Simpson", xlab="Value", col="yellow3")
hist(meta.red.comb$invsimp, breaks=20, main="Inverse Simpson", xlab="Value", col="yellow4")
dev.off()

plot(meta.red.comb[,c("richness","shannon","simpson","invsimp")])
```

## Trends in genus richness and diversity ~ Tanner breast stage (vagina + vulva samples)
```{r div-tb-trends, fig.width=12, fig.height=8, echo=FALSE}
## Boxplot of richness ~ Tanner breast
gg.rich.tb <- ggplot(subset(meta.red.comb, type=="girl" & tan.br.dr %in% c(2,3,4,5)), 
                     aes(y=richness, x=factor(tan.br.dr), fill=site))
box.rich.tb <- gg.rich.tb + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  xlab("Tanner breast stage") +
  ylab("Genus-level richness") +
  ggtitle("Richness") +
  theme_cust

## Boxplot of shannon ~ Tanner breast
gg.shan.tb <- ggplot(subset(meta.red.comb, type=="girl" & tan.br.dr %in% c(2,3,4,5)), 
                     aes(y=shannon, x=factor(tan.br.dr), fill=site))
box.shan.tb <- gg.shan.tb + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  xlab("Tanner breast stage") +
  ylab("Genus-level Shannon diversity") +
  ggtitle("Shannon") +
  theme_cust

## Boxplot of simpson ~ Tanner breast
gg.simp.tb <- ggplot(subset(meta.red.comb, type=="girl" & tan.br.dr %in% c(2,3,4,5)), 
                     aes(y=simpson, x=factor(tan.br.dr), fill=site))
box.simp.tb <- gg.simp.tb + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  xlab("Tanner breast stage") +
  ylab("Genus-level Simpson diversity") +
  ggtitle("Simpson") +
  theme_cust

## Boxplot of invsimpson ~ Tanner breast
gg.invsimp.tb <- ggplot(subset(meta.red.comb, type=="girl" & tan.br.dr %in% c(2,3,4,5)), 
                     aes(y=invsimp, x=factor(tan.br.dr), fill=site))
box.invsimp.tb <- gg.invsimp.tb + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  xlab("Tanner breast stage") +
  ylab("Genus-level Inverse Simpson diversity") +
  ggtitle("Inverse Simpson") +
  theme_cust

multiplot(box.rich.tb, box.shan.tb, box.simp.tb, box.invsimp.tb,
          layout=matrix(c(1,2,3,4), ncol=2, byrow=TRUE))
```

## Trends in genus richness and diversity ~ menarche status (vagina + vulva samples)
```{r div-ms-trends, fig.width=12, fig.height=8, echo=FALSE}
## Boxplot of richness ~ menarche status
gg.rich.ms <- ggplot(subset(meta.red.comb, men.stat %in% c("pre","post","mom.post")), 
                     aes(y=richness, x=men.stat, fill=site))
box.rich.ms <- gg.rich.ms + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  scale_x_discrete(limits=c("pre","post","mom.post"),
                   labels=c("Girl Pre", "Girl Post", "Mom")) +
  xlab("Menarche status") +
  ylab("Genus-level richness") +  
  ggtitle("Richness") +
  theme_cust

## Boxplot of shannon ~ menarche status
gg.shan.ms <- ggplot(subset(meta.red.comb, men.stat %in% c("pre","post","mom.post")), 
                     aes(y=shannon, x=men.stat, fill=site))
box.shan.ms <- gg.shan.ms + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  scale_x_discrete(limits=c("pre","post","mom.post"),
                   labels=c("Girl Pre", "Girl Post", "Mom")) +
  xlab("Menarche status") +
  ylab("Genus-level Shannon diversity") +
  ggtitle("Shannon") +
  theme_cust

## Boxplot of simpson ~ menarche status
gg.simp.ms <- ggplot(subset(meta.red.comb, men.stat %in% c("pre","post","mom.post")), 
                     aes(y=simpson, x=men.stat, fill=site))
box.simp.ms <- gg.simp.ms + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  scale_x_discrete(limits=c("pre","post","mom.post"),
                   labels=c("Girl Pre", "Girl Post", "Mom")) +
  xlab("Menarche status") +
  ylab("Genus-level Simpson diversity") +
  ggtitle("Simpson") +
  theme_cust

## Boxplot of shannon ~ menarche status
gg.invsimp.ms <- ggplot(subset(meta.red.comb, men.stat %in% c("pre","post","mom.post")), 
                     aes(y=invsimp, x=men.stat, fill=site))
box.invsimp.ms <- gg.invsimp.ms + 
  geom_boxplot(width=0.5, outlier.shape=1) +
  scale_fill_manual(values=col.site) +
  scale_x_discrete(limits=c("pre","post","mom.post"),
                   labels=c("Girl Pre", "Girl Post", "Mom")) +
  xlab("Menarche status") +
  ylab("Genus-level Inverse Simpson diversity") +
  ggtitle("Inverse Simpson") +
  theme_cust

multiplot(box.rich.ms, box.shan.ms, box.simp.ms, box.invsimp.ms,
          layout=matrix(c(1,2,3,4), ncol=2, byrow=TRUE))
```

***Trends in genus-level richness and Simpson’s diversity index of vaginal and vulvar microbiota with pubertal development and menarche status.*** Genus counts were subsampled at a depth of 2000 observations per sample and used to calculate richness (total number of genera observed) and Simpson's diversity index. In both cases all taxa that could not be classified to the genus level are combined in a single 'Other' category; richness and diversity are therefore underestimated. 415 samples are represented: 181 vaginal samples and 190 vulvar samples from 30 girls, and 44 vaginal samples from 23 mothers. (A) Genus richness in relation to Tanner breast stage (girls only) on the left; richness in relation to menarche status (girls and mothers) on the right. Vagina and vulva samples are represented by light pink and dark magenta coloring, respectively. (B) Simpson's diversity in relation to Tanner breast stage on left, and menarche status on right. Each box represents the interquartile range, the whiskers represent the upper and lower quartiles, the horizontal line represents the median, and open circles represent outliers.
```{r diversity, fig.width=8, fig.height=6, echo=FALSE}
fig.s9a1 <- box.rich.tb +
  ggtitle("A") +
  guides(fill=FALSE) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))
fig.s9a2 <- box.rich.ms +
  ggtitle("") +
  scale_fill_manual(values=col.site[c("vag","vul")], 
                    name="Site", labels=c("Vagina","Vulva")) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))

fig.s9b1 <- box.simp.tb +
  ggtitle("B") +
  guides(fill=FALSE) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))
fig.s9b2 <- box.simp.ms +
  ggtitle("") +
  scale_fill_manual(values=col.site[c("vag","vul")], 
                    name="Site", labels=c("Vagina","Vulva")) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))

# pdf("supplemental/diversity.pdf", width=8, height=6)
multiplot(fig.s9a1, fig.s9a2, fig.s9b1, fig.s9b2,
          layout=matrix(c(1,2,3,4), ncol=2, byrow=TRUE))
# dev.off()
```

## Trends in genus richness and diversity ~ vaginal pH (vagina samples only)
```{r div-ph-trends, fig.width=12, fig.height=8, echo=FALSE}
## Scatterplot of richness ~ pH (vagina only)
gg.rich.ph <- ggplot(subset(meta.red.comb, site=="vag"),
                     aes(y=richness, x=ph, color=type))
scat.rich.ph <- gg.rich.ph +
#   geom_jitter(position=position_jitter(width = 0.05, height = 0.01), size=3, alpha=0.7) +
  geom_point(size=3, alpha=0.7) +
  stat_smooth(alpha=0.2, size=1.5) +
  scale_color_manual(values=col.type.alpha70,
                     name="Sample Type",
                     labels=c("Girl","Mother")) +
  xlab("Vaginal pH") +
  ylab("Genus-level richness") +
  ggtitle("Richness") +
  theme_cust

## Scatterplot of shannon ~ pH (vagina only)
gg.shan.ph <- ggplot(subset(meta.red.comb, site=="vag"),
                     aes(y=shannon, x=ph, color=type))
scat.shan.ph <- gg.shan.ph +
  geom_point(size=3, alpha=0.7) +
  stat_smooth(alpha=0.2, size=1.5) +
  scale_color_manual(values=col.type.alpha70,
                     name="Sample Type",
                     labels=c("Girl","Mother")) +
  xlab("Vaginal pH") +
  ylab("Genus-level Shannon diversity") +
  ggtitle("Shannon") +
  theme_cust

## Scatterplot of simpson ~ pH (vagina only)
gg.simp.ph <- ggplot(subset(meta.red.comb, site=="vag"),
                     aes(y=simpson, x=ph, color=type))
scat.simp.ph <- gg.simp.ph +
  geom_point(size=3, alpha=0.7) +
  stat_smooth(alpha=0.2, size=1.5) +
  scale_color_manual(values=col.type.alpha70,
                     name="Sample Type",
                     labels=c("Girl","Mother")) +
  xlab("Vaginal pH") +
  ylab("Genus-level Simpson diversity") +
  ggtitle("Simpson") +
  theme_cust

## Scatterplot of invsimp ~ pH (vagina only)
gg.invsimp.ph <- ggplot(subset(meta.red.comb, site=="vag"),
                     aes(y=invsimp, x=ph, color=type))
scat.invsimp.ph <- gg.invsimp.ph +
  geom_point(size=3, alpha=0.7) +
  stat_smooth(alpha=0.2, size=1.5) +
  scale_color_manual(values=col.type.alpha70,
                     name="Sample Type",
                     labels=c("Girl","Mother")) +
  xlab("Vaginal pH") +
  ylab("Genus-level Inverse Simpson diversity") +
  ggtitle("Inverse Simpson") +
  theme_cust

multiplot(scat.rich.ph, scat.shan.ph, scat.simp.ph, scat.invsimp.ph,
          layout=matrix(c(1,2,3,4), ncol=2, byrow=TRUE))
```

***Genus-level richness, Simpson’s diversity index and vaginal pH.*** Each point represents a vaginal sample from either a girl or a mother. Vaginal pH was recorded for 120 samples from 24 girls and 37 samples from 21 mothers. (A) Genus-level richness plotted against vaginal pH. A locally weighted scatterplot smoothing (LOESS) function was applied separately to girl and mother data points, with 95% confidence intervals indicated by the light grey regions. (B) Genus-level Simpson's diversity index plotted against vaginal pH with LOESS curves for girls and mothers.
```{r ph-diversity, fig.width=8, fig.height=4, echo=FALSE}
fig.s10a <- scat.rich.ph +
  ggtitle("A") +
  guides(color=FALSE) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))
fig.s10b <- scat.simp.ph +
  ggtitle("B") +
  scale_color_manual(values=col.type.alpha70[c("girl","mom")], 
                    name="Sample Type", 
                    labels=c("Girl","Mother")) +
  theme(plot.title=element_text(size=22, hjust=-0.15, vjust=1.5))

# pdf("supplemental/ph-diversity.pdf", width=8, height=4)
multiplot(fig.s10a, fig.s10b, layout=matrix(c(1,1,1,2,2,2,2), ncol=7, byrow=TRUE))
# dev.off()
```

***
# Save R workspace
This will save the workspace (data) in two separate images: one named with today's date, in case you ever need to restore that version, and another with a non-dated name that can be easily loaded into subsequent analyses.

### Cleanup
```{r cleanup}
rm(meta.red, box.invsimp.ms, box.invsimp.tb, box.ph.hclust, box.ph.hclust2, box.rich.ms, box.rich.tb,
   box.shan.ms, box.shan.tb, box.simp.ms, box.simp.tb, fig.s10a, fig.s10b, fig.s9a1, fig.s9a2, fig.s9b1, fig.s9b2,
   gg.invsimp.ms, gg.invsimp.ph, gg.invsimp.tb, gg.ph.hclust, gg.ph.hclust2, gg.rich.ms, gg.rich.ms, gg.rich.tb,
   gg.shan.ms, gg.shan.ph, gg.shan.tb, gg.simp.ms, gg.simp.ph, gg.simp.tb, ord, scat.invsimp.ph, scat.rich.ph, 
   scat.shan.ph, scat.simp.ph, x)
```

```{r save-workspace}
save.image(paste("data-postproc/06-post-review-", Sys.Date(), ".RData", sep=""))
save.image(paste("data-postproc/06-post-review-last-run.RData", sep=""))
```
